+++ sdk/dart/lib/zcashname_sdk.dart
@@ -2,7 +2,7 @@
 library;
 
 export 'src/actions.dart' show ZnsAction;
-export 'src/crypto.dart' show verifyEd25519;
+export 'src/crypto.dart' show verifyEd25519, hashLeaf, hashInternal;
 export 'src/exceptions.dart'
     show InvalidNameException, InvalidAddressException, ZcashAddressKind;
 export 'src/models.dart';
@@ -11,7 +11,11 @@ export 'src/payload_validation.dart' show validatePayload, znsActions;
 export 'src/prepared_actions.dart';
 export 'src/rpc.dart' show ZnsRpcException;
 export 'src/validators.dart'
-    show isValidName, isValidUnifiedAddress, isValidTransparentAddress;
+    show
+        isValidName,
+        normalizeName,
+        isValidUnifiedAddress,
+        isValidTransparentAddress;
 export 'src/zip321.dart'
     show Zip321Parts, buildZcashUri, parseZip321Uri, toBase64Url, decodeBase64Url;
 export 'src/zns.dart' show ZNS;